refactor(keyring-api): src/utils/* -> src/api/
