+++ packages/keyring-api/src/api/index.ts
@@ -6,4 +6,5 @@ export * from './export';
 export * from './request';
 export * from './response';
 export * from './transaction';
+export * from './pagination';
 export type * from './keyring';
